Fix UI bug on icon animation during shared element transition
When the shared element transition was triggered from the VersionsActivity to the ApplicationsActivity, the RecyclerView would unbind all views, but only kept 5 recycled view (default value) and discarded all other views.
Side effect (during the animation) was that shared element transition wouldn't use the valid view as destination which resulted in a random ImageView (the rebinded view) of the RecyclerView to be hidden and the correct view would already show the icon.